pyproject.toml
[tool.poetry]
name = "harlequin-cassandra"
version = "0.1.0"
description = "A Harlequin adapter for Cassandra."
authors = ["Vadim Khitrin <vkhitrin@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "harlequin_cassandra", from = "src" },
]

[tool.poetry.plugins."harlequin.adapter"]
cassandra = "harlequin_cassandra:HarlequinCassandraAdapter"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
harlequin = "^1.7"
cassandra-driver = "^3.29.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
pytest = "^7.4.3"
mypy = "^1.7.0"
pre-commit = "^3.5.0"
importlib_metadata = { version = ">=4.6.0", python = "<3.10.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
select = ["A", "B", "E", "F", "I"]
target-version = "py38"

[tool.mypy]
python_version = "3.8"
files = [
    "src/**/*.py",
    "tests/**/*.py",
]
mypy_path = "src:stubs"

show_column_numbers = true

# show error messages from unrelated files
follow_imports = "normal"

# be strict
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
strict_optional = true

warn_return_any = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true

no_implicit_reexport = true
strict_equality = true
